seo: extract all page metadata to seo/pages
